fix: instruction create
