+++ packages/create-agent/src/config.ts
@@ -1,0 +1,1 @@
+export const STABLE_AXIOMKIT = {};
