+++ packages/sei/package.json
@@ -31,6 +31,7 @@
     "bs58": "^6.0.0",
     "ethers": "6.15.0",
     "viem": "^2.33.0",
-    "zod": "catalog:"
+    "zod": "catalog:",
+    "chalk": "catalog:"
   }
 }
